requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8

# AI and Text Generation
openai==1.12.0
python-dotenv==1.0.1

# Audio Processing
gTTS==2.5.0
ffmpeg-python==0.2.0

# Video Processing
opencv-python==4.9.0.80
moviepy==1.0.3
torch==2.2.0
torchvision==0.17.0
numpy==1.26.3
facenet-pytorch==2.5.3
librosa==0.10.1
scipy==1.12.0
tqdm==4.66.1

# Optional GPU support (uncomment if you have NVIDIA GPU)
# torch==2.2.0+cu118
# torchvision==0.17.0+cu118

# Utility
python-multipart==0.0.9
Pillow==10.2.0